refactor: split typst_gen.rs into focused sub-modules
Extract image, shape, and header/footer codegen from the monolithic
typst_gen.rs (1,307 lines) into three focused sub-modules:

- typst_gen_images.rs: image preprocessing (crop, format), floating images
- typst_gen_shapes.rs: shapes, shadows, polygons, gradients, border styles
- typst_gen_headers.rs: header/footer content, page setup for flow/table pages

typst_gen.rs is reduced from 1,307 to 763 lines (42% reduction).
GenCtx and page-level orchestration remain in the parent module.

Related: #138
