+++ crates/office2pdf/src/render/typst_gen.rs
@@ -1,7 +1,10 @@
 use std::fmt::Write;
+
+#[cfg(test)]
 use std::io::Cursor;
 
-use image::{GenericImageView, ImageFormat as RasterImageFormat};
+#[cfg(test)]
+use image::ImageFormat as RasterImageFormat;
 
 use crate::config::ConvertOptions;
 use crate::error::ConvertError;
@@ -16,18 +19,27 @@ use crate::ir::{
 };
 
 use self::diagrams::{generate_chart, generate_smartart};
+use self::headers::{write_flow_page_setup, write_table_page_setup};
+use self::images::{generate_floating_image, generate_image, preprocess_image_asset};
 use self::lists::{
     can_render_fixed_text_list_inline, common_text_style, generate_fixed_text_list, generate_list,
     write_common_text_settings, write_fixed_text_default_par_settings,
 };
+use self::shapes::{border_line_style_to_typst, generate_shape, write_gradient_fill};
 use self::tables::generate_table;
 use self::text::*;
 use super::font_context::FontSearchContext;
 
 #[path = "typst_gen_diagrams.rs"]
 mod diagrams;
+#[path = "typst_gen_headers.rs"]
+mod headers;
+#[path = "typst_gen_images.rs"]
+mod images;
 #[path = "typst_gen_lists.rs"]
 mod lists;
+#[path = "typst_gen_shapes.rs"]
+mod shapes;
 #[path = "typst_gen_tables.rs"]
 mod tables;
 #[path = "typst_gen_text.rs"]
@@ -89,55 +101,6 @@ impl GenCtx {
         let id = self.next_text_box_id;
         self.next_text_box_id += 1;
         id
-    }
-}
-
-fn raster_image_format(format: ImageFormat) -> Option<RasterImageFormat> {
-    match format {
-        ImageFormat::Png => Some(RasterImageFormat::Png),
-        ImageFormat::Jpeg => Some(RasterImageFormat::Jpeg),
-        ImageFormat::Gif => Some(RasterImageFormat::Gif),
-        ImageFormat::Bmp => Some(RasterImageFormat::Bmp),
-        ImageFormat::Tiff => Some(RasterImageFormat::Tiff),
-        ImageFormat::Svg => None,
-    }
-}
-
-fn crop_to_pixels(crop: ImageCrop, width: u32, height: u32) -> Option<(u32, u32, u32, u32)> {
-    let left = ((crop.left.clamp(0.0, 1.0) * width as f64).round() as u32).min(width);
-    let top = ((crop.top.clamp(0.0, 1.0) * height as f64).round() as u32).min(height);
-    let right = ((crop.right.clamp(0.0, 1.0) * width as f64).round() as u32).min(width);
-    let bottom = ((crop.bottom.clamp(0.0, 1.0) * height as f64).round() as u32).min(height);
-    if left + right >= width || top + bottom >= height {
-        return None;
-    }
-    Some((left, top, width - left - right, height - top - bottom))
-}
-
-fn preprocess_image_asset(image: &ImageData) -> (Vec<u8>, ImageFormat) {
-    let Some(crop) = image.crop.filter(|crop| !crop.is_empty()) else {
-        return (image.data.clone(), image.format);
-    };
-    let Some(raster_format) = raster_image_format(image.format) else {
-        return (image.data.clone(), image.format);
-    };
-    let Ok(decoded) = image::load_from_memory_with_format(&image.data, raster_format) else {
-        return (image.data.clone(), image.format);
-    };
-    let (width, height) = decoded.dimensions();
-    let Some((left, top, crop_width, crop_height)) = crop_to_pixels(crop, width, height) else {
-        return (image.data.clone(), image.format);
-    };
-
-    let cropped = decoded.crop_imm(left, top, crop_width, crop_height);
-    let mut encoded = Cursor::new(Vec::new());
-    if cropped
-        .write_to(&mut encoded, RasterImageFormat::Png)
-        .is_ok()
-    {
-        (encoded.into_inner(), ImageFormat::Png)
-    } else {
-        (image.data.clone(), image.format)
     }
 }
 
@@ -615,410 +578,6 @@ fn generate_fixed_text_box(
     Ok(())
 }
 
-fn generate_shape(out: &mut String, shape: &Shape, width: f64, height: f64) {
-    // Render shadow as offset duplicate before main shape
-    if let Some(shadow) = &shape.shadow {
-        write_shadow_shape(out, shape, width, height, shadow);
-    }
-
-    let has_rotation = shape.rotation_deg.is_some();
-    if let Some(deg) = shape.rotation_deg {
-        let _ = write!(out, "#rotate({}deg)[", format_f64(deg));
-    }
-
-    match &shape.kind {
-        ShapeKind::Rectangle => {
-            out.push_str("#rect(");
-            write_shape_params(out, shape, width, height);
-            out.push_str(")\n");
-        }
-        ShapeKind::Ellipse => {
-            out.push_str("#ellipse(");
-            write_shape_params(out, shape, width, height);
-            out.push_str(")\n");
-        }
-        ShapeKind::Line { x2, y2 } => {
-            out.push_str("#line(");
-            let _ = write!(
-                out,
-                "start: (0pt, 0pt), end: ({}pt, {}pt)",
-                format_f64(*x2),
-                format_f64(*y2),
-            );
-            write_shape_stroke(out, &shape.stroke);
-            out.push_str(")\n");
-        }
-        ShapeKind::RoundedRectangle { radius_fraction } => {
-            let radius = radius_fraction * width.min(height);
-            out.push_str("#rect(");
-            write_shape_params(out, shape, width, height);
-            let _ = write!(out, ", radius: {}pt", format_f64(radius));
-            out.push_str(")\n");
-        }
-        ShapeKind::Polygon { vertices } => {
-            write_polygon(out, shape, width, height, vertices);
-        }
-    }
-
-    if has_rotation {
-        out.push_str("]\n");
-    }
-}
-
-/// Render a shadow approximation as an offset duplicate shape with reduced opacity.
-fn write_shadow_shape(out: &mut String, shape: &Shape, width: f64, height: f64, shadow: &Shadow) {
-    let dir_rad = shadow.direction.to_radians();
-    let dx = shadow.distance * dir_rad.cos();
-    let dy = shadow.distance * dir_rad.sin();
-    let alpha = (shadow.opacity * 255.0).round() as u8;
-
-    let _ = write!(
-        out,
-        "#place(top + left, dx: {}pt, dy: {}pt)[",
-        format_f64(dx),
-        format_f64(dy),
-    );
-
-    match &shape.kind {
-        ShapeKind::Line { .. } => {
-            // Lines don't have meaningful shadows; skip
-            out.push_str("]\n");
-            return;
-        }
-        ShapeKind::Polygon { vertices } => {
-            // Shadow for polygon: duplicate polygon with shadow color
-            out.push_str("#polygon(");
-            write_polygon_vertices(out, width, height, vertices);
-            let _ = write!(
-                out,
-                ", fill: rgb({}, {}, {}, {})",
-                shadow.color.r, shadow.color.g, shadow.color.b, alpha,
-            );
-            out.push_str(")]\n");
-            return;
-        }
-        _ => {}
-    }
-    let shape_cmd = match &shape.kind {
-        ShapeKind::Rectangle => "#rect(",
-        ShapeKind::Ellipse => "#ellipse(",
-        ShapeKind::RoundedRectangle { radius_fraction } => {
-            let _ = writeln!(
-                out,
-                "#rect(width: {}pt, height: {}pt, radius: {}pt, fill: rgb({}, {}, {}, {}))]",
-                format_f64(width),
-                format_f64(height),
-                format_f64(radius_fraction * width.min(height)),
-                shadow.color.r,
-                shadow.color.g,
-                shadow.color.b,
-                alpha,
-            );
-            return;
-        }
-        // Line and Polygon are handled by early returns above; any future
-        // variants gracefully skip the shadow rather than panicking.
-        _ => {
-            out.push_str("]\n");
-            return;
-        }
-    };
-    out.push_str(shape_cmd);
-    let _ = write!(
-        out,
-        "width: {}pt, height: {}pt, fill: rgb({}, {}, {}, {})",
-        format_f64(width),
-        format_f64(height),
-        shadow.color.r,
-        shadow.color.g,
-        shadow.color.b,
-        alpha,
-    );
-    out.push_str(")]\n");
-}
-
-/// Write fill color, using rgb with 4 args when opacity is set, rgb with 3 args otherwise.
-fn write_fill_color(out: &mut String, fill: &Color, opacity: Option<f64>) {
-    if let Some(op) = opacity {
-        let alpha = (op * 255.0).round() as u8;
-        let _ = write!(
-            out,
-            ", fill: rgb({}, {}, {}, {})",
-            fill.r, fill.g, fill.b, alpha
-        );
-    } else {
-        let _ = write!(out, ", fill: rgb({}, {}, {})", fill.r, fill.g, fill.b);
-    }
-}
-
-fn write_shape_params(out: &mut String, shape: &Shape, width: f64, height: f64) {
-    let _ = write!(
-        out,
-        "width: {}pt, height: {}pt",
-        format_f64(width),
-        format_f64(height),
-    );
-    if let Some(gradient) = &shape.gradient_fill {
-        out.push_str(", fill: ");
-        write_gradient_fill(out, gradient);
-    } else if let Some(fill) = &shape.fill {
-        write_fill_color(out, fill, shape.opacity);
-    }
-    write_shape_stroke(out, &shape.stroke);
-}
-
-/// Write stroke parameter for shapes, handling dash patterns.
-fn write_shape_stroke(out: &mut String, stroke: &Option<BorderSide>) {
-    if let Some(stroke) = stroke {
-        match stroke.style {
-            BorderLineStyle::Solid | BorderLineStyle::None => {
-                let _ = write!(
-                    out,
-                    ", stroke: {}pt + rgb({}, {}, {})",
-                    format_f64(stroke.width),
-                    stroke.color.r,
-                    stroke.color.g,
-                    stroke.color.b,
-                );
-            }
-            _ => {
-                let _ = write!(
-                    out,
-                    ", stroke: (paint: rgb({}, {}, {}), thickness: {}pt, dash: \"{}\")",
-                    stroke.color.r,
-                    stroke.color.g,
-                    stroke.color.b,
-                    format_f64(stroke.width),
-                    border_line_style_to_typst(stroke.style),
-                );
-            }
-        }
-    }
-}
-
-/// Write polygon vertex coordinates scaled to actual dimensions.
-fn write_polygon_vertices(out: &mut String, width: f64, height: f64, vertices: &[(f64, f64)]) {
-    for (i, (vx, vy)) in vertices.iter().enumerate() {
-        if i > 0 {
-            out.push_str(", ");
-        }
-        let _ = write!(
-            out,
-            "({}pt, {}pt)",
-            format_f64(vx * width),
-            format_f64(vy * height),
-        );
-    }
-}
-
-/// Generate a Typst `#polygon(...)` for an arbitrary polygon shape.
-fn write_polygon(
-    out: &mut String,
-    shape: &Shape,
-    width: f64,
-    height: f64,
-    vertices: &[(f64, f64)],
-) {
-    out.push_str("#polygon(");
-    write_polygon_vertices(out, width, height, vertices);
-    if let Some(gradient) = &shape.gradient_fill {
-        out.push_str(", fill: ");
-        write_gradient_fill(out, gradient);
-    } else if let Some(fill) = &shape.fill {
-        write_fill_color(out, fill, shape.opacity);
-    }
-    write_shape_stroke(out, &shape.stroke);
-    out.push_str(")\n");
-}
-
-/// Write a Typst `gradient.linear(...)` expression.
-///
-/// Stops are sorted by position before rendering because Typst requires
-/// gradient stop offsets to be in monotonic (non-decreasing) order.
-/// The first stop is clamped to 0% and the last to 100% as Typst requires.
-fn write_gradient_fill(out: &mut String, gradient: &GradientFill) {
-    // Typst requires at least 2 stops for gradient.linear().
-    // Fall back to solid fill if fewer than 2 stops.
-    if gradient.stops.len() < 2 {
-        if let Some(stop) = gradient.stops.first() {
-            let _ = write!(
-                out,
-                "rgb({}, {}, {})",
-                stop.color.r, stop.color.g, stop.color.b,
-            );
-        }
-        return;
-    }
-    let mut sorted_stops = gradient.stops.clone();
-    sorted_stops.sort_by(|a, b| {
-        a.position
-            .partial_cmp(&b.position)
-            .unwrap_or(std::cmp::Ordering::Equal)
-    });
-    // Typst requires first stop at 0% and last stop at 100%.
-    if let Some(first) = sorted_stops.first_mut() {
-        first.position = 0.0;
-    }
-    if let Some(last) = sorted_stops.last_mut() {
-        last.position = 1.0;
-    }
-    out.push_str("gradient.linear(");
-    for (i, stop) in sorted_stops.iter().enumerate() {
-        if i > 0 {
-            out.push_str(", ");
-        }
-        let pos_pct = (stop.position * 100.0).round() as i64;
-        let _ = write!(
-            out,
-            "(rgb({}, {}, {}), {}%)",
-            stop.color.r, stop.color.g, stop.color.b, pos_pct,
-        );
-    }
-    if gradient.angle.abs() > 0.001 {
-        let _ = write!(out, ", angle: {}deg", format_f64(gradient.angle));
-    }
-    out.push(')');
-}
-
-fn write_page_setup(out: &mut String, size: &PageSize, margins: &Margins) {
-    let _ = writeln!(
-        out,
-        "#set page(width: {}pt, height: {}pt, margin: (top: {}pt, bottom: {}pt, left: {}pt, right: {}pt))",
-        format_f64(size.width),
-        format_f64(size.height),
-        format_f64(margins.top),
-        format_f64(margins.bottom),
-        format_f64(margins.left),
-        format_f64(margins.right),
-    );
-}
-
-/// Write the full page setup for a FlowPage, including optional header/footer.
-fn write_flow_page_setup(out: &mut String, page: &FlowPage, size: &PageSize) {
-    if page.header.is_none() && page.footer.is_none() {
-        write_page_setup(out, size, &page.margins);
-        return;
-    }
-
-    let _ = write!(
-        out,
-        "#set page(width: {}pt, height: {}pt, margin: (top: {}pt, bottom: {}pt, left: {}pt, right: {}pt)",
-        format_f64(size.width),
-        format_f64(size.height),
-        format_f64(page.margins.top),
-        format_f64(page.margins.bottom),
-        format_f64(page.margins.left),
-        format_f64(page.margins.right),
-    );
-
-    if let Some(header) = &page.header {
-        if hf_needs_context(header) {
-            out.push_str(", header: context [");
-        } else {
-            out.push_str(", header: [");
-        }
-        generate_hf_content(out, header);
-        out.push(']');
-    }
-
-    if let Some(footer) = &page.footer {
-        if hf_needs_context(footer) {
-            out.push_str(", footer: context [");
-        } else {
-            out.push_str(", footer: [");
-        }
-        generate_hf_content(out, footer);
-        out.push(']');
-    }
-
-    out.push_str(")\n");
-}
-
-/// Write the full page setup for a TablePage, including optional header/footer.
-fn write_table_page_setup(out: &mut String, page: &TablePage, size: &PageSize) {
-    if page.header.is_none() && page.footer.is_none() {
-        write_page_setup(out, size, &page.margins);
-        return;
-    }
-
-    let _ = write!(
-        out,
-        "#set page(width: {}pt, height: {}pt, margin: (top: {}pt, bottom: {}pt, left: {}pt, right: {}pt)",
-        format_f64(size.width),
-        format_f64(size.height),
-        format_f64(page.margins.top),
-        format_f64(page.margins.bottom),
-        format_f64(page.margins.left),
-        format_f64(page.margins.right),
-    );
-
-    if let Some(header) = &page.header {
-        if hf_needs_context(header) {
-            out.push_str(", header: context [");
-        } else {
-            out.push_str(", header: [");
-        }
-        generate_hf_content(out, header);
-        out.push(']');
-    }
-
-    if let Some(footer) = &page.footer {
-        if hf_needs_context(footer) {
-            out.push_str(", footer: context [");
-        } else {
-            out.push_str(", footer: [");
-        }
-        generate_hf_content(out, footer);
-        out.push(']');
-    }
-
-    out.push_str(")\n");
-}
-
-/// Check if a header/footer contains any context-dependent fields (page number or total pages).
-fn hf_needs_context(hf: &HeaderFooter) -> bool {
-    hf.paragraphs.iter().any(|p| {
-        p.elements
-            .iter()
-            .any(|e| matches!(e, HFInline::PageNumber | HFInline::TotalPages))
-    })
-}
-
-/// Generate inline content for a header or footer.
-fn generate_hf_content(out: &mut String, hf: &HeaderFooter) {
-    for (i, para) in hf.paragraphs.iter().enumerate() {
-        if i > 0 {
-            out.push_str("\\\n");
-        }
-        // Apply paragraph alignment if set
-        if let Some(align) = para.style.alignment {
-            let align_str = match align {
-                Alignment::Left => "left",
-                Alignment::Center => "center",
-                Alignment::Right => "right",
-                Alignment::Justify => "left",
-            };
-            let _ = write!(out, "#align({align_str})[");
-        }
-        for elem in &para.elements {
-            match elem {
-                HFInline::Run(run) => {
-                    generate_run(out, run);
-                }
-                HFInline::PageNumber => {
-                    out.push_str("#counter(page).display()");
-                }
-                HFInline::TotalPages => {
-                    out.push_str("#counter(page).final().first()");
-                }
-            }
-        }
-        if para.style.alignment.is_some() {
-            out.push(']');
-        }
-    }
-}
-
 fn generate_block(out: &mut String, block: &Block, ctx: &mut GenCtx) -> Result<(), ConvertError> {
     match block {
         Block::Paragraph(para) => generate_paragraph(out, para),
@@ -1072,109 +631,6 @@ fn format_insets(insets: &Insets) -> String {
         format_f64(insets.bottom),
         format_f64(insets.left),
     )
-}
-
-fn border_line_style_to_typst(style: BorderLineStyle) -> &'static str {
-    match style {
-        BorderLineStyle::Solid => "solid",
-        BorderLineStyle::Dashed => "dashed",
-        BorderLineStyle::Dotted => "dotted",
-        BorderLineStyle::DashDot => "dash-dotted",
-        BorderLineStyle::DashDotDot => "dash-dotted",
-        BorderLineStyle::Double => "dashed",
-        BorderLineStyle::None => "solid",
-    }
-}
-
-fn generate_image(out: &mut String, img: &ImageData, ctx: &mut GenCtx) {
-    let path = ctx.add_image(img);
-    out.push_str("#image(\"");
-    out.push_str(&path);
-    out.push('"');
-
-    if let Some(w) = img.width {
-        let _ = write!(out, ", width: {}pt", format_f64(w));
-    }
-    if let Some(h) = img.height {
-        let _ = write!(out, ", height: {}pt", format_f64(h));
-    }
-
-    out.push_str(")\n");
-}
-
-/// Generate Typst markup for a floating image.
-///
-/// Uses `#place()` for absolute positioning. The wrap mode determines how text
-/// interacts with the image:
-/// - Behind/InFront/None: `#place()` with no text wrapping
-/// - Square/Tight/TopAndBottom: `#place()` with `float: true` for best-effort text flow
-fn generate_floating_image(out: &mut String, fi: &FloatingImage, ctx: &mut GenCtx) {
-    let path = ctx.add_image(&fi.image);
-
-    match fi.wrap_mode {
-        WrapMode::TopAndBottom => {
-            // Emit a block-level image — text above and below only
-            out.push_str("#block(width: 100%)[\n");
-            let _ = write!(
-                out,
-                "  #place(top + left, dx: {}pt, dy: 0pt)[",
-                format_f64(fi.offset_x)
-            );
-            out.push_str("#image(\"");
-            out.push_str(&path);
-            out.push('"');
-            if let Some(w) = fi.image.width {
-                let _ = write!(out, ", width: {}pt", format_f64(w));
-            }
-            if let Some(h) = fi.image.height {
-                let _ = write!(out, ", height: {}pt", format_f64(h));
-            }
-            out.push_str(")]\n");
-            // Reserve vertical space equal to image height
-            if let Some(h) = fi.image.height {
-                let _ = writeln!(out, "  #v({}pt)", format_f64(h));
-            }
-            out.push_str("]\n");
-        }
-        WrapMode::Behind | WrapMode::InFront | WrapMode::None => {
-            // Place the image at absolute position, no text wrapping
-            let _ = write!(
-                out,
-                "#place(top + left, dx: {}pt, dy: {}pt)[",
-                format_f64(fi.offset_x),
-                format_f64(fi.offset_y)
-            );
-            out.push_str("#image(\"");
-            out.push_str(&path);
-            out.push('"');
-            if let Some(w) = fi.image.width {
-                let _ = write!(out, ", width: {}pt", format_f64(w));
-            }
-            if let Some(h) = fi.image.height {
-                let _ = write!(out, ", height: {}pt", format_f64(h));
-            }
-            out.push_str(")]\n");
-        }
-        WrapMode::Square | WrapMode::Tight => {
-            // Best-effort text wrapping: use #place with float: true
-            let _ = write!(
-                out,
-                "#place(top + left, dx: {}pt, dy: {}pt, float: true)[",
-                format_f64(fi.offset_x),
-                format_f64(fi.offset_y)
-            );
-            out.push_str("#image(\"");
-            out.push_str(&path);
-            out.push('"');
-            if let Some(w) = fi.image.width {
-                let _ = write!(out, ", width: {}pt", format_f64(w));
-            }
-            if let Some(h) = fi.image.height {
-                let _ = write!(out, ", height: {}pt", format_f64(h));
-            }
-            out.push_str(")]\n");
-        }
-    }
 }
 
 fn generate_floating_text_box(
