+++ crates/office2pdf/src/render/typst_gen_headers.rs
@@ -1,0 +1,142 @@
+use std::fmt::Write;
+
+use super::*;
+
+pub(super) fn write_page_setup(out: &mut String, size: &PageSize, margins: &Margins) {
+    let _ = writeln!(
+        out,
+        "#set page(width: {}pt, height: {}pt, margin: (top: {}pt, bottom: {}pt, left: {}pt, right: {}pt))",
+        format_f64(size.width),
+        format_f64(size.height),
+        format_f64(margins.top),
+        format_f64(margins.bottom),
+        format_f64(margins.left),
+        format_f64(margins.right),
+    );
+}
+
+/// Write the full page setup for a FlowPage, including optional header/footer.
+pub(super) fn write_flow_page_setup(out: &mut String, page: &FlowPage, size: &PageSize) {
+    if page.header.is_none() && page.footer.is_none() {
+        write_page_setup(out, size, &page.margins);
+        return;
+    }
+
+    let _ = write!(
+        out,
+        "#set page(width: {}pt, height: {}pt, margin: (top: {}pt, bottom: {}pt, left: {}pt, right: {}pt)",
+        format_f64(size.width),
+        format_f64(size.height),
+        format_f64(page.margins.top),
+        format_f64(page.margins.bottom),
+        format_f64(page.margins.left),
+        format_f64(page.margins.right),
+    );
+
+    if let Some(header) = &page.header {
+        if hf_needs_context(header) {
+            out.push_str(", header: context [");
+        } else {
+            out.push_str(", header: [");
+        }
+        generate_hf_content(out, header);
+        out.push(']');
+    }
+
+    if let Some(footer) = &page.footer {
+        if hf_needs_context(footer) {
+            out.push_str(", footer: context [");
+        } else {
+            out.push_str(", footer: [");
+        }
+        generate_hf_content(out, footer);
+        out.push(']');
+    }
+
+    out.push_str(")\n");
+}
+
+/// Write the full page setup for a TablePage, including optional header/footer.
+pub(super) fn write_table_page_setup(out: &mut String, page: &TablePage, size: &PageSize) {
+    if page.header.is_none() && page.footer.is_none() {
+        write_page_setup(out, size, &page.margins);
+        return;
+    }
+
+    let _ = write!(
+        out,
+        "#set page(width: {}pt, height: {}pt, margin: (top: {}pt, bottom: {}pt, left: {}pt, right: {}pt)",
+        format_f64(size.width),
+        format_f64(size.height),
+        format_f64(page.margins.top),
+        format_f64(page.margins.bottom),
+        format_f64(page.margins.left),
+        format_f64(page.margins.right),
+    );
+
+    if let Some(header) = &page.header {
+        if hf_needs_context(header) {
+            out.push_str(", header: context [");
+        } else {
+            out.push_str(", header: [");
+        }
+        generate_hf_content(out, header);
+        out.push(']');
+    }
+
+    if let Some(footer) = &page.footer {
+        if hf_needs_context(footer) {
+            out.push_str(", footer: context [");
+        } else {
+            out.push_str(", footer: [");
+        }
+        generate_hf_content(out, footer);
+        out.push(']');
+    }
+
+    out.push_str(")\n");
+}
+
+/// Check if a header/footer contains any context-dependent fields (page number or total pages).
+fn hf_needs_context(hf: &HeaderFooter) -> bool {
+    hf.paragraphs.iter().any(|p| {
+        p.elements
+            .iter()
+            .any(|e| matches!(e, HFInline::PageNumber | HFInline::TotalPages))
+    })
+}
+
+/// Generate inline content for a header or footer.
+fn generate_hf_content(out: &mut String, hf: &HeaderFooter) {
+    for (i, para) in hf.paragraphs.iter().enumerate() {
+        if i > 0 {
+            out.push_str("\\\n");
+        }
+        // Apply paragraph alignment if set
+        if let Some(align) = para.style.alignment {
+            let align_str = match align {
+                Alignment::Left => "left",
+                Alignment::Center => "center",
+                Alignment::Right => "right",
+                Alignment::Justify => "left",
+            };
+            let _ = write!(out, "#align({align_str})[");
+        }
+        for elem in &para.elements {
+            match elem {
+                HFInline::Run(run) => {
+                    generate_run(out, run);
+                }
+                HFInline::PageNumber => {
+                    out.push_str("#counter(page).display()");
+                }
+                HFInline::TotalPages => {
+                    out.push_str("#counter(page).final().first()");
+                }
+            }
+        }
+        if para.style.alignment.is_some() {
+            out.push(']');
+        }
+    }
+}
