+++ crates/office2pdf/src/render/typst_gen_images.rs
@@ -1,0 +1,150 @@
+use std::fmt::Write;
+use std::io::Cursor;
+
+use image::{GenericImageView, ImageFormat as RasterImageFormat};
+
+use super::*;
+
+fn raster_image_format(format: ImageFormat) -> Option<RasterImageFormat> {
+    match format {
+        ImageFormat::Png => Some(RasterImageFormat::Png),
+        ImageFormat::Jpeg => Some(RasterImageFormat::Jpeg),
+        ImageFormat::Gif => Some(RasterImageFormat::Gif),
+        ImageFormat::Bmp => Some(RasterImageFormat::Bmp),
+        ImageFormat::Tiff => Some(RasterImageFormat::Tiff),
+        ImageFormat::Svg => None,
+    }
+}
+
+pub(super) fn crop_to_pixels(
+    crop: ImageCrop,
+    width: u32,
+    height: u32,
+) -> Option<(u32, u32, u32, u32)> {
+    let left = ((crop.left.clamp(0.0, 1.0) * width as f64).round() as u32).min(width);
+    let top = ((crop.top.clamp(0.0, 1.0) * height as f64).round() as u32).min(height);
+    let right = ((crop.right.clamp(0.0, 1.0) * width as f64).round() as u32).min(width);
+    let bottom = ((crop.bottom.clamp(0.0, 1.0) * height as f64).round() as u32).min(height);
+    if left + right >= width || top + bottom >= height {
+        return None;
+    }
+    Some((left, top, width - left - right, height - top - bottom))
+}
+
+pub(super) fn preprocess_image_asset(image: &ImageData) -> (Vec<u8>, ImageFormat) {
+    let Some(crop) = image.crop.filter(|crop| !crop.is_empty()) else {
+        return (image.data.clone(), image.format);
+    };
+    let Some(raster_format) = raster_image_format(image.format) else {
+        return (image.data.clone(), image.format);
+    };
+    let Ok(decoded) = image::load_from_memory_with_format(&image.data, raster_format) else {
+        return (image.data.clone(), image.format);
+    };
+    let (width, height) = decoded.dimensions();
+    let Some((left, top, crop_width, crop_height)) = crop_to_pixels(crop, width, height) else {
+        return (image.data.clone(), image.format);
+    };
+
+    let cropped = decoded.crop_imm(left, top, crop_width, crop_height);
+    let mut encoded = Cursor::new(Vec::new());
+    if cropped
+        .write_to(&mut encoded, RasterImageFormat::Png)
+        .is_ok()
+    {
+        (encoded.into_inner(), ImageFormat::Png)
+    } else {
+        (image.data.clone(), image.format)
+    }
+}
+
+pub(super) fn generate_image(out: &mut String, img: &ImageData, ctx: &mut GenCtx) {
+    let path = ctx.add_image(img);
+    out.push_str("#image(\"");
+    out.push_str(&path);
+    out.push('"');
+
+    if let Some(w) = img.width {
+        let _ = write!(out, ", width: {}pt", format_f64(w));
+    }
+    if let Some(h) = img.height {
+        let _ = write!(out, ", height: {}pt", format_f64(h));
+    }
+
+    out.push_str(")\n");
+}
+
+/// Generate Typst markup for a floating image.
+///
+/// Uses `#place()` for absolute positioning. The wrap mode determines how text
+/// interacts with the image:
+/// - Behind/InFront/None: `#place()` with no text wrapping
+/// - Square/Tight/TopAndBottom: `#place()` with `float: true` for best-effort text flow
+pub(super) fn generate_floating_image(out: &mut String, fi: &FloatingImage, ctx: &mut GenCtx) {
+    let path = ctx.add_image(&fi.image);
+
+    match fi.wrap_mode {
+        WrapMode::TopAndBottom => {
+            // Emit a block-level image — text above and below only
+            out.push_str("#block(width: 100%)[\n");
+            let _ = write!(
+                out,
+                "  #place(top + left, dx: {}pt, dy: 0pt)[",
+                format_f64(fi.offset_x)
+            );
+            out.push_str("#image(\"");
+            out.push_str(&path);
+            out.push('"');
+            if let Some(w) = fi.image.width {
+                let _ = write!(out, ", width: {}pt", format_f64(w));
+            }
+            if let Some(h) = fi.image.height {
+                let _ = write!(out, ", height: {}pt", format_f64(h));
+            }
+            out.push_str(")]\n");
+            // Reserve vertical space equal to image height
+            if let Some(h) = fi.image.height {
+                let _ = writeln!(out, "  #v({}pt)", format_f64(h));
+            }
+            out.push_str("]\n");
+        }
+        WrapMode::Behind | WrapMode::InFront | WrapMode::None => {
+            // Place the image at absolute position, no text wrapping
+            let _ = write!(
+                out,
+                "#place(top + left, dx: {}pt, dy: {}pt)[",
+                format_f64(fi.offset_x),
+                format_f64(fi.offset_y)
+            );
+            out.push_str("#image(\"");
+            out.push_str(&path);
+            out.push('"');
+            if let Some(w) = fi.image.width {
+                let _ = write!(out, ", width: {}pt", format_f64(w));
+            }
+            if let Some(h) = fi.image.height {
+                let _ = write!(out, ", height: {}pt", format_f64(h));
+            }
+            out.push_str(")]\n");
+        }
+        WrapMode::Square | WrapMode::Tight => {
+            // Best-effort text wrapping: use #place with float: true
+            let _ = write!(
+                out,
+                "#place(top + left, dx: {}pt, dy: {}pt, float: true)[",
+                format_f64(fi.offset_x),
+                format_f64(fi.offset_y)
+            );
+            out.push_str("#image(\"");
+            out.push_str(&path);
+            out.push('"');
+            if let Some(w) = fi.image.width {
+                let _ = write!(out, ", width: {}pt", format_f64(w));
+            }
+            if let Some(h) = fi.image.height {
+                let _ = write!(out, ", height: {}pt", format_f64(h));
+            }
+            out.push_str(")]\n");
+        }
+    }
+}
