+++ crates/office2pdf/src/render/typst_gen_shapes.rs
@@ -1,0 +1,280 @@
+use std::fmt::Write;
+
+use super::*;
+
+pub(super) fn generate_shape(out: &mut String, shape: &Shape, width: f64, height: f64) {
+    // Render shadow as offset duplicate before main shape
+    if let Some(shadow) = &shape.shadow {
+        write_shadow_shape(out, shape, width, height, shadow);
+    }
+
+    let has_rotation = shape.rotation_deg.is_some();
+    if let Some(deg) = shape.rotation_deg {
+        let _ = write!(out, "#rotate({}deg)[", format_f64(deg));
+    }
+
+    match &shape.kind {
+        ShapeKind::Rectangle => {
+            out.push_str("#rect(");
+            write_shape_params(out, shape, width, height);
+            out.push_str(")\n");
+        }
+        ShapeKind::Ellipse => {
+            out.push_str("#ellipse(");
+            write_shape_params(out, shape, width, height);
+            out.push_str(")\n");
+        }
+        ShapeKind::Line { x2, y2 } => {
+            out.push_str("#line(");
+            let _ = write!(
+                out,
+                "start: (0pt, 0pt), end: ({}pt, {}pt)",
+                format_f64(*x2),
+                format_f64(*y2),
+            );
+            write_shape_stroke(out, &shape.stroke);
+            out.push_str(")\n");
+        }
+        ShapeKind::RoundedRectangle { radius_fraction } => {
+            let radius = radius_fraction * width.min(height);
+            out.push_str("#rect(");
+            write_shape_params(out, shape, width, height);
+            let _ = write!(out, ", radius: {}pt", format_f64(radius));
+            out.push_str(")\n");
+        }
+        ShapeKind::Polygon { vertices } => {
+            write_polygon(out, shape, width, height, vertices);
+        }
+    }
+
+    if has_rotation {
+        out.push_str("]\n");
+    }
+}
+
+/// Render a shadow approximation as an offset duplicate shape with reduced opacity.
+fn write_shadow_shape(out: &mut String, shape: &Shape, width: f64, height: f64, shadow: &Shadow) {
+    let dir_rad = shadow.direction.to_radians();
+    let dx = shadow.distance * dir_rad.cos();
+    let dy = shadow.distance * dir_rad.sin();
+    let alpha = (shadow.opacity * 255.0).round() as u8;
+
+    let _ = write!(
+        out,
+        "#place(top + left, dx: {}pt, dy: {}pt)[",
+        format_f64(dx),
+        format_f64(dy),
+    );
+
+    match &shape.kind {
+        ShapeKind::Line { .. } => {
+            // Lines don't have meaningful shadows; skip
+            out.push_str("]\n");
+            return;
+        }
+        ShapeKind::Polygon { vertices } => {
+            // Shadow for polygon: duplicate polygon with shadow color
+            out.push_str("#polygon(");
+            write_polygon_vertices(out, width, height, vertices);
+            let _ = write!(
+                out,
+                ", fill: rgb({}, {}, {}, {})",
+                shadow.color.r, shadow.color.g, shadow.color.b, alpha,
+            );
+            out.push_str(")]\n");
+            return;
+        }
+        _ => {}
+    }
+    let shape_cmd = match &shape.kind {
+        ShapeKind::Rectangle => "#rect(",
+        ShapeKind::Ellipse => "#ellipse(",
+        ShapeKind::RoundedRectangle { radius_fraction } => {
+            let _ = writeln!(
+                out,
+                "#rect(width: {}pt, height: {}pt, radius: {}pt, fill: rgb({}, {}, {}, {}))]",
+                format_f64(width),
+                format_f64(height),
+                format_f64(radius_fraction * width.min(height)),
+                shadow.color.r,
+                shadow.color.g,
+                shadow.color.b,
+                alpha,
+            );
+            return;
+        }
+        // Line and Polygon are handled by early returns above; any future
+        // variants gracefully skip the shadow rather than panicking.
+        _ => {
+            out.push_str("]\n");
+            return;
+        }
+    };
+    out.push_str(shape_cmd);
+    let _ = write!(
+        out,
+        "width: {}pt, height: {}pt, fill: rgb({}, {}, {}, {})",
+        format_f64(width),
+        format_f64(height),
+        shadow.color.r,
+        shadow.color.g,
+        shadow.color.b,
+        alpha,
+    );
+    out.push_str(")]\n");
+}
+
+/// Write fill color, using rgb with 4 args when opacity is set, rgb with 3 args otherwise.
+fn write_fill_color(out: &mut String, fill: &Color, opacity: Option<f64>) {
+    if let Some(op) = opacity {
+        let alpha = (op * 255.0).round() as u8;
+        let _ = write!(
+            out,
+            ", fill: rgb({}, {}, {}, {})",
+            fill.r, fill.g, fill.b, alpha
+        );
+    } else {
+        let _ = write!(out, ", fill: rgb({}, {}, {})", fill.r, fill.g, fill.b);
+    }
+}
+
+fn write_shape_params(out: &mut String, shape: &Shape, width: f64, height: f64) {
+    let _ = write!(
+        out,
+        "width: {}pt, height: {}pt",
+        format_f64(width),
+        format_f64(height),
+    );
+    if let Some(gradient) = &shape.gradient_fill {
+        out.push_str(", fill: ");
+        write_gradient_fill(out, gradient);
+    } else if let Some(fill) = &shape.fill {
+        write_fill_color(out, fill, shape.opacity);
+    }
+    write_shape_stroke(out, &shape.stroke);
+}
+
+/// Write stroke parameter for shapes, handling dash patterns.
+pub(super) fn write_shape_stroke(out: &mut String, stroke: &Option<BorderSide>) {
+    if let Some(stroke) = stroke {
+        match stroke.style {
+            BorderLineStyle::Solid | BorderLineStyle::None => {
+                let _ = write!(
+                    out,
+                    ", stroke: {}pt + rgb({}, {}, {})",
+                    format_f64(stroke.width),
+                    stroke.color.r,
+                    stroke.color.g,
+                    stroke.color.b,
+                );
+            }
+            _ => {
+                let _ = write!(
+                    out,
+                    ", stroke: (paint: rgb({}, {}, {}), thickness: {}pt, dash: \"{}\")",
+                    stroke.color.r,
+                    stroke.color.g,
+                    stroke.color.b,
+                    format_f64(stroke.width),
+                    border_line_style_to_typst(stroke.style),
+                );
+            }
+        }
+    }
+}
+
+/// Write polygon vertex coordinates scaled to actual dimensions.
+fn write_polygon_vertices(out: &mut String, width: f64, height: f64, vertices: &[(f64, f64)]) {
+    for (i, (vx, vy)) in vertices.iter().enumerate() {
+        if i > 0 {
+            out.push_str(", ");
+        }
+        let _ = write!(
+            out,
+            "({}pt, {}pt)",
+            format_f64(vx * width),
+            format_f64(vy * height),
+        );
+    }
+}
+
+/// Generate a Typst `#polygon(...)` for an arbitrary polygon shape.
+fn write_polygon(
+    out: &mut String,
+    shape: &Shape,
+    width: f64,
+    height: f64,
+    vertices: &[(f64, f64)],
+) {
+    out.push_str("#polygon(");
+    write_polygon_vertices(out, width, height, vertices);
+    if let Some(gradient) = &shape.gradient_fill {
+        out.push_str(", fill: ");
+        write_gradient_fill(out, gradient);
+    } else if let Some(fill) = &shape.fill {
+        write_fill_color(out, fill, shape.opacity);
+    }
+    write_shape_stroke(out, &shape.stroke);
+    out.push_str(")\n");
+}
+
+/// Write a Typst `gradient.linear(...)` expression.
+///
+/// Stops are sorted by position before rendering because Typst requires
+/// gradient stop offsets to be in monotonic (non-decreasing) order.
+/// The first stop is clamped to 0% and the last to 100% as Typst requires.
+pub(super) fn write_gradient_fill(out: &mut String, gradient: &GradientFill) {
+    // Typst requires at least 2 stops for gradient.linear().
+    // Fall back to solid fill if fewer than 2 stops.
+    if gradient.stops.len() < 2 {
+        if let Some(stop) = gradient.stops.first() {
+            let _ = write!(
+                out,
+                "rgb({}, {}, {})",
+                stop.color.r, stop.color.g, stop.color.b,
+            );
+        }
+        return;
+    }
+    let mut sorted_stops = gradient.stops.clone();
+    sorted_stops.sort_by(|a, b| {
+        a.position
+            .partial_cmp(&b.position)
+            .unwrap_or(std::cmp::Ordering::Equal)
+    });
+    // Typst requires first stop at 0% and last stop at 100%.
+    if let Some(first) = sorted_stops.first_mut() {
+        first.position = 0.0;
+    }
+    if let Some(last) = sorted_stops.last_mut() {
+        last.position = 1.0;
+    }
+    out.push_str("gradient.linear(");
+    for (i, stop) in sorted_stops.iter().enumerate() {
+        if i > 0 {
+            out.push_str(", ");
+        }
+        let pos_pct = (stop.position * 100.0).round() as i64;
+        let _ = write!(
+            out,
+            "(rgb({}, {}, {}), {}%)",
+            stop.color.r, stop.color.g, stop.color.b, pos_pct,
+        );
+    }
+    if gradient.angle.abs() > 0.001 {
+        let _ = write!(out, ", angle: {}deg", format_f64(gradient.angle));
+    }
+    out.push(')');
+}
+
+pub(super) fn border_line_style_to_typst(style: BorderLineStyle) -> &'static str {
+    match style {
+        BorderLineStyle::Solid => "solid",
+        BorderLineStyle::Dashed => "dashed",
+        BorderLineStyle::Dotted => "dotted",
+        BorderLineStyle::DashDot => "dash-dotted",
+        BorderLineStyle::DashDotDot => "dash-dotted",
+        BorderLineStyle::Double => "dashed",
+        BorderLineStyle::None => "solid",
+    }
+}
